requirements.txt
aiohttp==3.11.11
beautifulsoup4==4.12.3
bs4==0.0.2
cryptography==44.0.0
dnspython==2.7.0
multidict==6.1.0
urllib3==2.3.0
webtech==1.3.4
